+++ src/constants/graph.js
@@ -1,7 +1,7 @@
 export const NODE_BACKGROUND = '#adefd1' // 'adefd1' // '03a9f4' // '2B7CE9',
 export const NODE_BORDER = '#011e41' // 'D2E5FF',
 export const NODE_BORDER_WIDTH = 1
-export const CLICK_NODE_BACKGROUND = '#abd6df'
+export const CLICK_NODE_BACKGROUND = '#ffed00' // '#abd6df'
 export const HIGHLIGHT_NODE_BACKGROUND = '#ffef00'
 export const HIGHLIGHT_NODE_BORDER = '#009688'
 export const HOVER_NODE_BACKGROUND = '#f2f2f2'